Revert "LPS-187086 Rename ThreadLocalScopeContextScopeChecker to ThreadLocalScopeChecker, and extract ScopeContext to ThreadLocalScopeContext"
This reverts commit 7f1dd683a2a85e5379cf604d6d39e11a02dced20.
